tableを表示するコンポーネントのディレクリ構成を変更
