+++ frontend/app/components/parts/trainingRecordTable/TrainingRecordTable.tsx
@@ -1,12 +1,11 @@
 import type { TrainingRecordWithName } from "~/type/training_record";
 import { trainingRecordDelete } from "~/apiActions/TrainingRecord";
-import TrainingRecordListTable from "./TrainingRecordListTable";
+import BaseTrainingRecordTable from "./BaseTrainingRecordTable";
 
 type TrainingRecordProps = {
-  date: string;
   trainingRecord: TrainingRecordWithName[];
   currentPage: number;
-  getTrainingRecord: (date: string) => Promise<
+  getTrainingRecord: () => Promise<
     | {
         success: boolean;
         data: any;
@@ -24,9 +23,8 @@ type TrainingRecordProps = {
   setCurrentPage: (page: number) => void;
 };
 
-// 日付ごとの筋トレ実績一覧を生成する関数コンポーネント
-const TrainingRecordTableByDate = ({
-  date,
+// 筋トレ実績一覧を生成する関数コンポーネント
+const TrainingRecordListTable = ({
   trainingRecord,
   currentPage,
   getTrainingRecord,
@@ -38,7 +36,7 @@ const TrainingRecordTableByDate = ({
     const response = await trainingRecordDelete(id);
     if (response.success) {
       alert("削除が完了しました。");
-      const result = await getTrainingRecord(date);
+      const result = await getTrainingRecord();
       if (result.success) {
         setTrainingRecord(result.data);
       }
@@ -46,8 +44,9 @@ const TrainingRecordTableByDate = ({
       alert(`削除できませんでした。\n\n${response.error}`);
     }
   };
+
   return (
-    <TrainingRecordListTable
+    <BaseTrainingRecordTable
       trainingRecord={trainingRecord}
       currentPage={currentPage}
       setCurrentPage={setCurrentPage}
@@ -55,4 +54,4 @@ const TrainingRecordTableByDate = ({
     />
   );
 };
-export default TrainingRecordTableByDate;
+export default TrainingRecordListTable;
